feat: allow slowdown (negative fast forward multipliers)
- not the most understandable thing for now, but works as expected, negative fast forward multiplier -> relative slowdown based on the basefpstarget in a 1/X format

+++ src/platform/wasm/main.h
@@ -34,7 +34,7 @@ struct mEmscriptenRenderer {
 
 	// persistent options for the core at runtime, limited
 	// subset of mCoreOptions and custom functionality
-	int fastForwardMultiplier;
+	double fastForwardMultiplier;
 	int frameSkip;
 	int baseFpsTarget;
 	int rewindBufferCapacity;
